Add a function getPodDependencies for InterpretDependency operation
