+++ pkg/resourceinterpreter/customized/declarative/luavm/kube.go
@@ -28,6 +28,9 @@ const (
 //   - function accuratePodRequirements(pod) requirements
 //     accurate total resource requirements for pod. Example:
 //     requirements = kube.accuratePodRequirements(pod)
+//   - function getPodDependencies(podTemplate, namespace) dependencies
+//     get total dependencies from podTemplate and namespace. Example:
+//     dependencies = kube.getPodDependencies(podTemplate, namespace)
 func KubeLoader(ls *lua.LState) int {
 	mod := ls.SetFuncs(ls.NewTable(), kubeFuncs)
 	ls.Push(mod)
@@ -37,6 +40,7 @@ func KubeLoader(ls *lua.LState) int {
 var kubeFuncs = map[string]lua.LGFunction{
 	"resourceAdd":             resourceAdd,
 	"accuratePodRequirements": accuratePodRequirements,
+	"getPodDependencies":      getPodDependencies,
 }
 
 func resourceAdd(ls *lua.LState) int {
@@ -78,6 +82,35 @@ func accuratePodRequirements(ls *lua.LState) int {
 	return 1
 }
 
+func getPodDependencies(ls *lua.LState) int {
+	n := ls.GetTop()
+	if n != 2 {
+		ls.RaiseError("getPodDependencies only accepts two argument")
+		return 0
+	}
+
+	podTemplate := ls.CheckTable(1)
+	namespace := checkNamespace(ls, 2)
+
+	template := &corev1.PodTemplateSpec{}
+	err := ConvertLuaResultInto(podTemplate, template)
+	if err != nil {
+		ls.RaiseError("fail to convert lua value %#v to corev1.PodTemplateSpec: %v", podTemplate, err)
+		return 0
+	}
+
+	pod := helper.GeneratePodFromTemplateAndNamespace(template, namespace)
+	dependencies, _ := helper.GetDependenciesFromPodTemplate(pod)
+	retValue, err := decodeValue(ls, dependencies)
+	if err != nil {
+		ls.RaiseError("fail to convert %#v to Lua value: %v", dependencies, err)
+		return 0
+	}
+
+	ls.Push(retValue)
+	return 1
+}
+
 func checkResourceQuantity(ls *lua.LState, n int) resource.Quantity {
 	v := ls.Get(n)
 	switch typ := v.Type(); typ {
@@ -89,5 +122,18 @@ func checkResourceQuantity(ls *lua.LState, n int) resource.Quantity {
 	default:
 		ls.TypeError(n, lua.LTString)
 		return resource.Quantity{}
+	}
+}
+
+func checkNamespace(ls *lua.LState, n int) string {
+	v := ls.Get(n)
+	switch typ := v.Type(); typ {
+	case lua.LTNil:
+		return "default"
+	case lua.LTString, lua.LTNumber:
+		return ls.CheckString(n)
+	default:
+		ls.TypeError(n, lua.LTString)
+		return ""
 	}
 }
