+++ pkg/resourceinterpreter/default/native/dependencies.go
@@ -8,7 +8,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/util/sets"
 
 	configv1alpha1 "github.com/karmada-io/karmada/pkg/apis/config/v1alpha1"
 	"github.com/karmada-io/karmada/pkg/util"
@@ -40,7 +39,7 @@ func getDeploymentDependencies(object *unstructured.Unstructured) ([]configv1alp
 		return nil, err
 	}
 
-	return getDependenciesFromPodTemplate(podObj)
+	return helper.GetDependenciesFromPodTemplate(podObj)
 }
 
 func getJobDependencies(object *unstructured.Unstructured) ([]configv1alpha1.DependentObjectReference, error) {
@@ -55,7 +54,7 @@ func getJobDependencies(object *unstructured.Unstructured) ([]configv1alpha1.Dep
 		return nil, err
 	}
 
-	return getDependenciesFromPodTemplate(podObj)
+	return helper.GetDependenciesFromPodTemplate(podObj)
 }
 
 func getCronJobDependencies(object *unstructured.Unstructured) ([]configv1alpha1.DependentObjectReference, error) {
@@ -70,7 +69,7 @@ func getCronJobDependencies(object *unstructured.Unstructured) ([]configv1alpha1
 		return nil, err
 	}
 
-	return getDependenciesFromPodTemplate(podObj)
+	return helper.GetDependenciesFromPodTemplate(podObj)
 }
 
 func getPodDependencies(object *unstructured.Unstructured) ([]configv1alpha1.DependentObjectReference, error) {
@@ -80,7 +79,7 @@ func getPodDependencies(object *unstructured.Unstructured) ([]configv1alpha1.Dep
 		return nil, fmt.Errorf("failed to convert Pod from unstructured object: %v", err)
 	}
 
-	return getDependenciesFromPodTemplate(podObj)
+	return helper.GetDependenciesFromPodTemplate(podObj)
 }
 
 func getDaemonSetDependencies(object *unstructured.Unstructured) ([]configv1alpha1.DependentObjectReference, error) {
@@ -95,7 +94,7 @@ func getDaemonSetDependencies(object *unstructured.Unstructured) ([]configv1alph
 		return nil, err
 	}
 
-	return getDependenciesFromPodTemplate(podObj)
+	return helper.GetDependenciesFromPodTemplate(podObj)
 }
 
 func getStatefulSetDependencies(object *unstructured.Unstructured) ([]configv1alpha1.DependentObjectReference, error) {
@@ -110,89 +109,5 @@ func getStatefulSetDependencies(object *unstructured.Unstructured) ([]configv1al
 		return nil, err
 	}
 
-	return getDependenciesFromPodTemplate(podObj)
-}
-
-func getDependenciesFromPodTemplate(podObj *corev1.Pod) ([]configv1alpha1.DependentObjectReference, error) {
-	dependentConfigMaps := getConfigMapNames(podObj)
-	dependentSecrets := getSecretNames(podObj)
-	dependentSas := getServiceAccountNames(podObj)
-	dependentPVCs := getPVCNames(podObj)
-	var dependentObjectRefs []configv1alpha1.DependentObjectReference
-	for cm := range dependentConfigMaps {
-		dependentObjectRefs = append(dependentObjectRefs, configv1alpha1.DependentObjectReference{
-			APIVersion: "v1",
-			Kind:       "ConfigMap",
-			Namespace:  podObj.Namespace,
-			Name:       cm,
-		})
-	}
-
-	for secret := range dependentSecrets {
-		dependentObjectRefs = append(dependentObjectRefs, configv1alpha1.DependentObjectReference{
-			APIVersion: "v1",
-			Kind:       "Secret",
-			Namespace:  podObj.Namespace,
-			Name:       secret,
-		})
-	}
-	for sa := range dependentSas {
-		dependentObjectRefs = append(dependentObjectRefs, configv1alpha1.DependentObjectReference{
-			APIVersion: "v1",
-			Kind:       "ServiceAccount",
-			Namespace:  podObj.Namespace,
-			Name:       sa,
-		})
-	}
-
-	for pvc := range dependentPVCs {
-		dependentObjectRefs = append(dependentObjectRefs, configv1alpha1.DependentObjectReference{
-			APIVersion: "v1",
-			Kind:       "PersistentVolumeClaim",
-			Namespace:  podObj.Namespace,
-			Name:       pvc,
-		})
-	}
-
-	return dependentObjectRefs, nil
-}
-
-func getSecretNames(pod *corev1.Pod) sets.Set[string] {
-	result := sets.New[string]()
-	lifted.VisitPodSecretNames(pod, func(name string) bool {
-		result.Insert(name)
-		return true
-	})
-	return result
-}
-
-func getServiceAccountNames(pod *corev1.Pod) sets.Set[string] {
-	result := sets.New[string]()
-	if pod.Spec.ServiceAccountName != "" && pod.Spec.ServiceAccountName != "default" {
-		result.Insert(pod.Spec.ServiceAccountName)
-	}
-	return result
-}
-
-func getConfigMapNames(pod *corev1.Pod) sets.Set[string] {
-	result := sets.New[string]()
-	lifted.VisitPodConfigmapNames(pod, func(name string) bool {
-		result.Insert(name)
-		return true
-	})
-	return result
-}
-
-func getPVCNames(pod *corev1.Pod) sets.Set[string] {
-	result := sets.New[string]()
-	for i := range pod.Spec.Volumes {
-		volume := pod.Spec.Volumes[i]
-		if volume.PersistentVolumeClaim != nil {
-			claimName := volume.PersistentVolumeClaim.ClaimName
-			if len(claimName) != 0 {
-				result.Insert(claimName)
-			}
-		}
-	}
-	return result
+	return helper.GetDependenciesFromPodTemplate(podObj)
 }
